Added the "auto_redirect" directive.
If a prefix location ends with a slash, and requests are processed with "*_pass"
or "api" directives, a custom behavior occurs, called an "auto redirect".

Such a redirect occurs when a request arrives with a URI that has no trailing
slash but otherwise matches the prefix exactly; then, a permanent 301 code
redirect is returned pointing to the requested URI with the slash appended.

The new directive provides immediate control over this:

 - on - enable auto redirect
 - off - disable auto redirect
 - default - doesn't override the default for a location,
             which depends on the presence of the above directives.

The "default" value is actually the default for the directive.