QUIC: added debug logging of stream creation.
Currently, it is difficult to associate a stream connection number with its
parent connection.  The typical case is to identify the QUIC connection number
given a user-visible URI (which occurs in the request stream).

The patch adds a debug log message that reports stream creation in the stream
log and also shows the parent connection number.