QUIC: fixed closing of streams processed during the zrtt phase.
Existing checks in ngx_quic_reset_stream() and ngx_quic_shutdown_stream()
assume that closing streams before the handshake is complete is only possible
for client streams (which were in initial state and can be closed silently).

With early data, it is possible for streams to be created, used, and even
closed before the handshake is complete.  Silently closing them results in the
FIN flag not being sent to the client.  Some clients (e.g. Chrome) mark such a
connection as an error on close: the stream was incomplete even though all
bytes were sent.

The client behavior could be different if the connection was closed at QUIC
level (idle timeout and no error) and at HTTP3 level (keepalive timeout and
NGX_HTTP_V3_ERR_NO_ERROR error message).

The fix is to silently close streams only in client mode: the zrtt is not used
by the client and the code works as intended.

The issue was introduced by 31bad89f3430 (Angie 1.4.0).