Tests: guarded ssl_stapling.t for old OpenSSL versions.
Stapling requires the certificate status TLS extension.  It is implemented
in OpenSSL 0.9.8h and available if TLS extensions support is explicitly
configured, and by default since 0.9.8j.

To properly test if stapling is available, OpenSSL version is checked,
and SNI availability is checked to ensure TLS extensions support is
compiled in.