Tests: explicit Valgrind support.
Valgrind logging is done to a separate file, as it is not able to
follow stderr redirection within nginx or append to a file without
corrupting it.  Further, Valgrind logging seems to interfere with
error suppression in tests, and catches various startup errors and
warnings, so the log is additionally filtered.

Since startup under Valgrind can be really slow, timeout in waitforfile()
was changed to 10 seconds.

Prodded by Robert Mueller.