Mail: fixed EXTERNAL to be accepted only if enabled.
As originally implemented in 6774:bcb107bb89cd, it wasn't possible to
disable the EXTERNAL authentication method: it was always accepted
(but not advertised unless enabled).  It is, however, believed that
it is better to reject attempts to use the disabled method, hence in
6869:b2915d99ee8d an attempt was made to address this.  This attempt
was insufficient though: it was still possible to use the method as long
as initial SASL response was used.

With this patch both challenge-response and initial response forms are
disabled.  Additionally, initial response handling for the PLAIN
authentication is removed from ngx_mail_auth_parse(), for consistency
and to don't provoke such bugs.
(grafted from 6c54c8107baf00692cd82a0a8ed251af4b9224ad)