Core: Command line option to display build environment.
Currently, there is no easy way to know which features have been detected by
the configure script for the given binary.  Only configure arguments are
available, which does not provide enough information about the environment.
Reproducing the build environment or examining build logs for a specific binary
can be a tricky and time-consuming process.

The patch converts all definitions from the ngx_auto_config.h file to
human-readable text, which is output on demand at the command line.