Core: added "-m" and "-M" switches to show modules.
While looking at the configure output, it can be hard to understand what is
built and what is not because some modules are built by default, some are
disabled, and some are enabled (maybe even multiple times).

The "-m" shows builtin  modules and does not require configuration test.
The "-M" parses configuration and shows all modules, including external.

The list provided by this option is definitive and leaves no doubt.