Core: error logging rate limiting.
With this change, error logging to files can be rate-limited with
the "rate=" parameter.  The parameter specifies allowed log messages
rate to a particular file (per worker), in messages per second (m/s).
By default, "rate=1000m/s" is used.

Rate limiting is implemented using the "leaky bucket" method, similarly
to the limit_req module.

Maximum burst size is set to the number of log messages per second
for each severity level, so "error" messages are logged even if the
rate limit is hit by "info" messages (but not vice versa).  When the
limit is reached for a particular level, the "too many log messages,
limiting" message is logged at this level.

If debug logging is enabled, either for the particular log file or for
the particular connection, rate limiting is not used.
(grafted from 067c9c527d077c10ad4557b750c5384c1cbe9272)