Configure: makefile generator optimizations.
The generator makes multiple sed(1) calls, even when not really needed.
The code is thrashed with multiple copies of regular expressions making
it hard to read.

Most sed calls are replaced with shell builtins and moved into functions, each
performing clear task, making the rest of the code easier to read. Functions
save results to variables instead of echoing.

The most often and useless case: replacing slash with configured directory
separator (which is also slash on UNIX platform) using sed(1) is optimized.