Creation of the PID file made atomic.
Previously, it was possible that the PID file is already created but empty,
which made it harder to handle by external programs.

Before the process of the PID file creation looked like this:

  fd = open("angie.pid", O_CREAT|O_TRUNC);
  write(fd, pid);
  close(fd);

So, there was a possibility that the PID file can be read by and an external
process between open() and write() calls.

Now the PID number is written to an intermediate temporary file and then, after
the file is closed and all the data synced to file system, the temporary file
is renamed to the actual PID file name:

  fd = open("angie.pid.tmp", O_CREAT|O_TRUNC);
  write(fd, pid);
  close(fd);
  rename("angie.pid.tmp", "angie.pid");

This ensures that when "angie.pid" is appeared on the file system, it already
contains the PID number, which, inter alia, should make "binary_upgrade.t" test
more stable.