ACME: fixed crashes on configurations with many clients.
Previously, worker processes could crash on configurations with 5 or more
ACME clients.

When the client array grows, it may copy its elements to a new allocation.
As a result, any previously stored pointers to a particular element will
incorrectly point to memory that isn't properly initialized.  As a result,
an attempt to access key or certificate variables of this client will crash
a worker process with a NULL pointer dereference.

The fix is to allocate client structures separately and store only pointers
to them in the array.

--HG--
branch : 1.6