Tests: added new Test::Nginx method run_tests().
This method is designed to run all test cases in a test file.
It also allows you to run a specific test case.
To use it, the test must be slightly rewritten - individual tests
must be grouped into functions (subtests), and the order in which
subtests are executed should not be significant.

Usage:

    my %test_cases = (
        'tc1: test case 1' => \&tc1,
        'tc2: test case 2' => \&tc2,
        'tc3: test case 3' => \&tc3,
    );

    $t->plan(scalar keys %test_cases);

    $t->run_tests(\%test_cases);

To run a "tc2" test case add TEST_ANGIE_TC="tc2: test case 2" before prove.