Tests: improved tests of protection against additional USR2.
Previously, it waited for a "changing binary" message in the log before sending
the second USR2 signal, but there are two matching log messages:

  [notice]: signal 12 (SIGUSR2) received from PID, changing binary
  [notice]: changing binary

The first appears when the signal is received, and the second when it's actually
processed.

If the USR2 signal is sent before the second message, it will cause the first
log message to appear again, breaking the test that is waiting for the message
about ignoring the signal, which only appears if the binary upgrade process has
already started.

But it's still a good case to test.  So now we send USR2 after the first message
and another USR2 after the second message, and only then the test waits for the
ignored signal message.