Configure: run autotests by external means.
The intended use is the cross-compilation, when produced binaries cannot
be run on the build host.  The patch provides a path to script that wraps
autotests execution.  The script may hardcode expected result or execute
it on some remote target/virtual machine.

The new configure option is added: --test-runner=</path/to/script>.
An example script is provided in auto/crossrun.sh.

Reviewed at https://rb.tech.wbsrv.ru/r/558/