Separate methods for handling connections and load balancing.
This simplifies keepalive, sticky and load-balancing modules.

For sticky, it is no longer necessary to call the original balancer on success.
Previously it was necessary to call the keepalive module.

The ngx_event_connect_peer() no longer returns NGX_DONE.  The connection reuse
status is already stored in the peer.cached flag.