ACME: minor optimization of shared memory usage.
Previously, a shared memory region was allocated for the key authorization
string, even when it was unnecessary.

When Angie handles an HTTP validation challenge, worker processes respond to
the challenge by sending a key authorization string taken from shared memory.
When hooks are used, worker processes don't respond to the challenge, so
the shared memory region is not necessary.