HTTP/3: proxying with dynamic table fixes.
Previously, access to a dynamic table (triggered by sending inserts upstream)
resulted in a null pointer dereference because a stub connection was used
without an initialized HTTP/3 session.

An additional test for the current connection ensures that we distinguish
between initial request processing (when the connection is alive) and cache
processing without a connection to the upstream.  In the latter case, we
additionally initialize session object and configuration contexts so that
parsing can perform dynamic table checks and access configuration.

This allows us to remove extra checks in http_v3_parse.c and be closer to the
original nginx code.

The real problem is that caching HTTP/3 with dynamic table currently does not
work at all: to parse the cached response, we need the state of the encoder
stream that generates inserts into the table. And since the encoder stream is
an unrelated nginx connection object, it is not present in the cache. So
parsing fails saying that the stream is blocked.  This leads to 'invalid
headers from upstream' error and new request to upstream. And this requires to
reset v3parse state on request reinit handler in proxy module.

The patch disables dynamic table usage when proxying with cache enabled.