ACME: changed logic of "no valid domain name defined" message.
This patch changes the logic of the "no valid domain name defined" error
message. Previously, Angie would fail at startup with this error
message if an ACME client was defined but not referenced in any of the
server blocks. Now, in this case, Angie continues to run, and the ACME
client gets disabled. However, Angie will still fail with this error
message if no valid (i.e. ACME-compliant) domain name is found in
any of the server blocks the ACME client is referenced in.

Reviewed at https://rb.tech.wbsrv.ru/r/707/