Fixed configuration inheritance with multiple certificates.
Multiple certificates are enabled only when building using Tongsuo with NTLS.

The problem manifests itself in configurations like this:

http {

    server {
        listen ...

        proxy_ssl_certificate $arg_cert.example.com.crt;
        proxy_ssl_certificate_key $arg_cert.example.com.key;

        location /first {
            proxy_pass https://...
        }

        location /second {
            proxy_pass https://...;
        }
    }
}

The certificates are using variables and the ssl context is created in the
first location, and preserved into the server, so the second location
inherits existing ssl context.

The problem is that ssl context does not have loaded certificates (they are
using variables and should be loaded in runtime). In runtime, the value
of plcf->upstream.ssl_certificate_values is missing in the second location,
because it is not calculated in proxy_set_ssl() - the function exists early
since the ssl context already exists.

The fix is to treat ssl_certificate_values and ssl_certificate_key_values
normally as a part of configuration and merge them as usual, so they can
be always used in runtime.

The block that tests various certificates misconfiguration and compiles
variables is moved from the proxy_set_ssl() function into the
proxy_ssl_compile_certificates() function and executed during normal merge
process. This also required to move ssl_passwords merge before the latter.

The final bit of the fix is the change to ngx_ssl_preserve_passwords()
that accounts for special case of zero-items passwords array. In case
of multiple certificates such situation is possible and the
ngx_ssl_cache_pkey_password_callback() may access this, leading to
Address Sanitizer hit.