Support for Multipath TCP on Linux.
With this change, Multipath TCP can be enabled on Linux with
"listen ... multipath".  The listen ... multipath option is
supported in the http, stream and mail modules.

Note that to enable Multipath TCP on Linux, you must create
a socket with the IPPROTO_MPTCP protocol explicitly specified,
and it is not possible to modify an existing socket.  To allow
a transition with minimal impact on client connections, when
the multipath option is changed in the configuration, SO_REUSEPORT
is set on the old socket and the new socket is opened with
IPPROTO_MPTCP.

Note that this creates a race window, and connection requests
assigned to the old socket will be lost.  In particular, this
may affect binary upgrades if the WINCH signal is used to preserve
the old master process.

Requires Linux kernel 5.6 or later.  Note, however, that some of
the socket options may not be supported with Multipath TCP, or
may only be supported in newer kernels.  In particular, TCP_NODELAY
is only supported with kernel version 5.17 or later.

Based on patches by Maxime Dourov and Anthony Doeraene.
(grafted from 30e0c97dc26015e9afbc84dff2e451f67835bbf5)