Tests: improved checking of master process PIDs.
Previously, we considered all processes with the word "master" in the name
to be master processes. But sometimes these processes turned out to be
workers that had not yet been renamed, and the test would erroneously fail.

Now we look for all processes with the word "master" in the name and
consider only those that have child processes to be masters.

Additionally, cleaning of command execution results from leading
and trailing whitespace characters has been added.