Upstream zone: fixed peer refcount locking.
The update of the "zombies" field in the "ngx_http_upstream_rr_peers_t"
must be done under the lock.

This is not true for the ngx_http_upstream_rr_peer_unref().  Also, the
function itself must be called while holding the peer lock.

The same issue is fixed in the stream module.